+++ src/develop.rs
@@ -1,3 +1,4 @@
+use crate::build_options::CargoOptions;
 use crate::BuildOptions;
 use crate::PlatformTag;
 use crate::PythonInterpreter;
@@ -32,14 +33,15 @@ pub fn develop(
         interpreter: vec![python.clone()],
         find_interpreter: false,
         bindings,
-        manifest_path: Some(manifest_file.to_path_buf()),
         out: Some(wheel_dir.path().to_path_buf()),
         skip_auditwheel: false,
         zig: false,
-        target: None,
-        cargo_extra_args,
-        rustc_extra_args,
         universal2: false,
+        cargo: CargoOptions {
+            target: None,
+            manifest_path: Some(manifest_file.to_path_buf()),
+            ..Default::default()
+        },
     };
 
     let build_context = build_options.into_build_context(release, strip, true)?;